writing_practice/requirements.txt
streamlit==1.31.1
openai==0.28.1
python-dotenv==1.0.1
pytesseract==0.3.10
Pillow==10.2.0
pyyaml==6.0.1
requests==2.31.0
gTTS==2.5.4
